+++ packages/web-components/keen-slider-auto-switch/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brytdesigns/web-component-keen-slider-auto-switch
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/web-component-keen-slider@1.0.9
+  - @brytdesigns/web-component-utils@0.1.11
+
 ## 1.0.5
 
 ### Patch Changes