+++ shopify/packages/hybrid-cart-components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/shopify-hybrid-cart-components
 
+## 0.2.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/web-component-utils@0.1.11
+
 ## 0.2.28
 
 ### Patch Changes